refactor: Rename CreateTutorProfile to have a more explicit title
